check

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/domain/Member.java
@@ -13,7 +13,7 @@ public class Member implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long memberId;
+    private Integer memberId;
     private  String firstName;
     private String lastName;
